test(ai-codegen): keep generator regressions with module
