+++ packages/tools/openapi-generator/test/OpenApiPatch.test.ts
@@ -1,64 +1,13 @@
-import * as Config from "@effect/ai-codegen/Config"
-import type { DiscoveredProvider } from "@effect/ai-codegen/Discovery"
-import * as Generator from "@effect/ai-codegen/Generator"
-import * as OpenApiGenerator from "@effect/openapi-generator/OpenApiGenerator"
 import * as OpenApiPatch from "@effect/openapi-generator/OpenApiPatch"
 import * as NodeServices from "@effect/platform-node/NodeServices"
 import { assert, describe, it } from "@effect/vitest"
 import * as Effect from "effect/Effect"
 import * as Exit from "effect/Exit"
-import * as Layer from "effect/Layer"
 import * as Path from "effect/Path"
 
 const testLayer = NodeServices.layer
-const generatorLayer = Generator.layer.pipe(
-  Layer.provide(Layer.succeed(OpenApiGenerator.OpenApiGenerator, {
-    generate: (spec) => Effect.succeed(JSON.stringify(spec))
-  })),
-  Layer.provideMerge(testLayer)
-)
-
-const generateWithPatch = (patch: string) =>
-  Effect.gen(function*() {
-    const provider: DiscoveredProvider = {
-      name: "test",
-      packagePath: "/providers/test",
-      config: new Config.CodegenConfig({
-        spec: "spec.json",
-        output: "Generated.ts",
-        patches: [patch]
-      }),
-      specSource: Config.SpecSource.File("spec.json"),
-      outputPath: "Generated.ts"
-    }
-    const generator = yield* Generator.CodeGenerator
-    const output = yield* generator.generate(provider, {
-      openapi: "3.1.0",
-      info: { title: "Before", version: "1.0.0" },
-      paths: {}
-    })
-    return JSON.parse(output)
-  }).pipe(Effect.provide(generatorLayer))
 
 describe("OpenApiPatch", () => {
-  describe("CodeGenerator patch inputs", () => {
-    it.effect("accepts inline patches with leading whitespace", () =>
-      Effect.gen(function*() {
-        const result = yield* generateWithPatch(
-          " \n[{\"op\":\"replace\",\"path\":\"/info/title\",\"value\":\"After\"}]"
-        )
-        assert.strictEqual(result.info.title, "After")
-      }))
-
-    it.effect("accepts absolute paths with leading whitespace", () =>
-      Effect.gen(function*() {
-        const path = yield* Path.Path
-        const patch = path.join(import.meta.dirname, "fixtures/patches/valid-replace.json")
-        const result = yield* generateWithPatch(` ${patch}`)
-        assert.strictEqual(result.info.title, "Updated API Title")
-      }).pipe(Effect.provide(testLayer)))
-  })
-
   describe("parsePatchInput", () => {
     describe("inline JSON", () => {
       it.effect("parses valid inline JSON patch", () =>
